Create Scatter
Scatter plot is one of the visualization techniques in Pandas for representing data from a DataFrame. To make a simple scatterplot, we can use the plot. scatter() function. The plot() function can also be used to generate a scatter plot, as well as a bar graph, plot box, histogram, and plot bar in Pandas.